% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.outcome-response-priming.R
\name{process_orp}
\alias{process_orp}
\title{Process Outcome Reponse Priming Task data file}
\usage{
process_orp(path, local = FALSE, p)
}
\arguments{
\item{path}{Path to data file}

\item{local}{TRUE for locally saved JSON (default=FALSE)}

\item{p}{Participant identifier}
}
\value{
Data frame
}
\description{
\code{process_orp()} processes data files generated by the Outcome Response Priming task
(\url{https://earcanal.github.io/outcome-response-priming/}).
}
\examples{
\dontrun{
orp_df <- expand.grid(token = participants$token) \%>\%
 mutate(path=paste('data/', token, '_finished/outcome-response-priming-task-results.json', sep=''))
orp_df <- orp_df \%>\%
group_by(token) \%>\%
 do(., process_orp(.$path, .$token)) \%>\%
 ungroup \%>\%
 left_join(participants, by='token')
}
}
\keyword{expfactory}
\keyword{outcome}
\keyword{priming}
\keyword{response}
